Move shared types to model from gui
